Break apart RssServer class
