+++ packages/shared/src/types/rss.types.ts
@@ -1,4 +1,4 @@
-import type {AsyncResult} from '@conductor/shared/src/types/results.types';
+import type {AsyncResult, Result} from '@conductor/shared/src/types/results.types';
 import {z} from 'zod';
 
 export type RssFeedProviderType = 'local' | 'superfeedr';
@@ -26,4 +26,18 @@ export interface RssFeedItem {
   readonly link: string;
   readonly pubDate: Date;
   readonly content?: string;
+}
+
+export interface RssFeedSubscription {
+  readonly feedUrl: string;
+  readonly callbackUrl: string;
+}
+
+export interface RssFeedManager {
+  getFeed(args: {feedId: string}): RssFeed | null;
+  addFeed(args: {feed: RssFeed}): Result<void>;
+  updateFeed(args: {feedId: string; items: RssFeedItem[]}): AsyncResult<void>;
+  subscribe(args: {feedUrl: string; callbackUrl: string}): void;
+  unsubscribe(args: {feedUrl: string}): void;
+  getSubscriptions(args: {feedUrl: string}): Set<RssFeedSubscription> | null;
 }
